+++ src/m_menu.c
@@ -3253,6 +3253,11 @@ static const char *bobbing_styles[] = {
   "Vanilla", "Inv. Vanilla", "Alpha", "Inv. Alpha", "Smooth", "Inv. Smooth", "Quake", NULL
 };
 
+// [Nugget]
+static const char *show_berserk_strings[] = {
+  "None", "Ammo count", "Arms", "Both", NULL
+};
+
 setup_menu_t weap_settings2[] =
 {
   // [Nugget] These three
@@ -3285,8 +3290,8 @@ setup_menu_t weap_settings2[] =
   {"Enable Recoil Pitch", S_YESNO, m_null, M_X,
    M_Y + weap2_recoilpitch*M_SPC, {"weapon_recoilpitch"}},
    
-  {"Show Berserk when using Fist", S_YESNO, m_null, M_X, // [Nugget]
-   M_Y + weap2_berserk*M_SPC, {"show_berserk"}},
+  {"Show Berserk availability in", S_CHOICE, m_null, M_X, // [Nugget]
+   M_Y + weap2_berserk*M_SPC, {"show_berserk"}, 0, NULL, show_berserk_strings},
      
   {"Squat Weapon Down On Impact", S_YESNO|S_STRICT, m_null, M_X, // [Nugget]
    M_Y + weap2_squat*M_SPC, {"weaponsquat"}},